Rewrite and fix timeout issue
The timeout issue was due to invalid construction of response.
The responses made to GETPDU command in SNMP must be exact and directed
toward requestee's ID. I fixed that certain issue, however there is A
LOT to be done in here.

Also, I cleaned the code (made methods static as being methods wouldn't
do them much anyways) and documented a bit. Took the liberty of removing
extra libraries and removing threading as it is a simple liner which I
believe any user can pull off.

From here, it should be pretty easy to implement the methods used from
delimitry and to go based on the examples set by him, to implement other
response types and to study how to properly decode the OID for the
client not to throw back jibber on the terminal.

I added a simple config and suggest you use 'black' from now on.